@@ -178,7 +178,7 @@ internal distributed actor ClusterSingletonBoss<Act: ClusterSingleton>: ClusterS
             // Finally, release the singleton -- it should not have been refered to strongly by anyone else,
             // causing the instance to be released. TODO: we could assert that we have released it soon after here (it's ID must be resigned).
             self.actorSystem.releaseWellKnownActorID(instance.id)
-            await self.updateSingleton(nil)
+            self.updateSingleton(nil)
         }
     }
 
@@ -464,7 +464,7 @@ extension ClusterSingletonBoss {
             {
                 assert(
                     singleton.id.node == selfNode,
-                    "Target singleton node and targetNode were not the same! TargetNode: \(targetNode)," +
+                    "Target singleton node and targetNode were not the same! TargetNode: \(targetNode?.debugDescription ?? "UNKNOWN TARGET NODE")," +
                         " singleton.id.node: \(singleton.id.node)"
                 )
                 return try await singleton.actorSystem.localCall(
@@ -518,7 +518,7 @@ extension ClusterSingletonBoss {
 
             assert(
                 singleton.id.node == selfNode,
-                "Target singleton node and targetNode were not the same! TargetNode: \(targetNode)," +
+                "Target singleton node and targetNode were not the same! TargetNode: \(targetNode?.debugDescription ?? "UNKNOWN TARGET NODE")," +
                     " singleton.id.node: \(singleton.id.node)"
             )
             return try await singleton.actorSystem.localCallVoid(
